MDEV-19784 [2/4] errmsg: change %d to %zu for size specifiers to match size_t

+++ sql/handler.cc
@@ -403,8 +403,7 @@ handler *get_ha_partition(partition_info *part_info)
   }
   else
   {
-    my_error(ER_OUTOFMEMORY, MYF(ME_FATAL), 
-             static_cast<int>(sizeof(ha_partition)));
+    my_error(ER_OUTOFMEMORY, MYF(ME_FATAL), sizeof(ha_partition));
   }
   DBUG_RETURN(((handler*) partition));
 }
